Stub out controller methods
- Parse configuration as booleans
- Store configuration on plugin instance
- Move DataJsonController to separate file
